requirements.txt
yfinance>=0.2.42
pymc>=5.10.0
pandas>=2.0.0
numpy>=1.24.0
arviz>=0.18.0
xlrd>=2.0.1
statsmodels>=0.14.0